+++ src/ui/trackedit/TrackEdit.cpp
@@ -71,8 +71,6 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     wxStaticText *      NuStaticText;
     wxStaticText *      GeStaticText;
     wxStaticText *      YeStaticText;
-    wxPanel *           PicturePanel;
-    wxPanel *           MBrainzPanel;
     wxStaticText *      MBAlbumStaticText;
     wxStaticLine *      MBrainzStaticLine;
 
@@ -91,7 +89,7 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     //wxDialog( parent, wxID_ANY, _( "Songs Editor" ), wxDefaultPosition, wxSize( 625, 440 ), wxDEFAULT_DIALOG_STYLE | wxRESIZE_BORDER )
     Create( parent, wxID_ANY, _( "Songs Editor" ), WindowPos, WindowSize, wxDEFAULT_DIALOG_STYLE | wxRESIZE_BORDER | wxMAXIMIZE_BOX );
 
-//  this->SetSizeHints( wxDefaultSize, wxDefaultSize );
+    //this->SetSizeHints( wxDefaultSize, wxDefaultSize );
 
     wxBoxSizer * MainSizer = new wxBoxSizer( wxVERTICAL );
 
@@ -136,7 +134,7 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     //
     // Details
     //
-    wxPanel * DetailPanel = new wxPanel( m_MainNotebook, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxTAB_TRAVERSAL );
+    m_DetailPanel = new wxPanel( m_MainNotebook, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxTAB_TRAVERSAL );
     wxSizer * MainDetailSizer = new wxBoxSizer( wxVERTICAL );
 
     wxFlexGridSizer * DataFlexSizer = new wxFlexGridSizer( 3, 0, 0 );
@@ -145,139 +143,139 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     DataFlexSizer->SetNonFlexibleGrowMode( wxFLEX_GROWMODE_SPECIFIED );
 
     // Title
-    m_TiCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_TiCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_TiCopyButton->SetToolTip( _( "Copy the title to all the tracks you are editing" ) );
     DataFlexSizer->Add( m_TiCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
-    TiStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Title")), wxDefaultPosition, wxDefaultSize, 0 );
+    TiStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Title")), wxDefaultPosition, wxDefaultSize, 0 );
     TiStaticText->Wrap( -1 );
     DataFlexSizer->Add( TiStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT|wxTOP|wxRIGHT, 5 );
 
-    m_TitleTextCtrl = new wxTextCtrl( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
+    m_TitleTextCtrl = new wxTextCtrl( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
     DataFlexSizer->Add( m_TitleTextCtrl, 0, wxEXPAND|wxTOP|wxRIGHT, 5 );
 
     // Artist
-    m_ArCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_ArCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_ArCopyButton->SetToolTip( _( "Copy the artist name to all the tracks you are editing" ) );
     DataFlexSizer->Add( m_ArCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
-    ArStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Artist")), wxDefaultPosition, wxDefaultSize, 0 );
+    ArStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Artist")), wxDefaultPosition, wxDefaultSize, 0 );
     ArStaticText->Wrap( -1 );
     DataFlexSizer->Add( ArStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT|wxTOP|wxRIGHT, 5 );
 
     wxArrayString DummyArray;
-    m_ArtistComboBox = new wxComboBox( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
+    m_ArtistComboBox = new wxComboBox( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
     DataFlexSizer->Add( m_ArtistComboBox, 1, wxEXPAND|wxTOP|wxRIGHT, 5 );
 
     // Album Artist
-    m_AACopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_AACopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_AACopyButton->SetToolTip( _( "Copy the album artist name to all the tracks you are editing" ) );
     DataFlexSizer->Add( m_AACopyButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
-    wxStaticText * AAStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("A. Artist")), wxDefaultPosition, wxDefaultSize, 0 );
+    wxStaticText * AAStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("A. Artist")), wxDefaultPosition, wxDefaultSize, 0 );
     AAStaticText->SetToolTip( _( "shows the album artist of the track" ) );
     AAStaticText->Wrap( -1 );
     DataFlexSizer->Add( AAStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT|wxTOP|wxRIGHT, 5 );
 
-    m_AlbumArtistComboBox = new wxComboBox( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
+    m_AlbumArtistComboBox = new wxComboBox( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
     DataFlexSizer->Add( m_AlbumArtistComboBox, 1, wxEXPAND|wxTOP|wxRIGHT, 5 );
 
     // Album
-    m_AlCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_AlCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_AlCopyButton->SetToolTip( _( "Copy the album name to all the tracks you are editing" ) );
     DataFlexSizer->Add( m_AlCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
-    AlStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Album")), wxDefaultPosition, wxDefaultSize, 0 );
+    AlStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Album")), wxDefaultPosition, wxDefaultSize, 0 );
     AlStaticText->Wrap( -1 );
     DataFlexSizer->Add( AlStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT|wxTOP|wxRIGHT, 5 );
 
-    m_AlbumComboBox = new wxComboBox( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
+    m_AlbumComboBox = new wxComboBox( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
     DataFlexSizer->Add( m_AlbumComboBox, 1, wxEXPAND|wxTOP|wxRIGHT, 5 );
 
     // Compositor
-    m_CoCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_CoCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_CoCopyButton->SetToolTip( _( "Copy the composer to all the tracks you are editing" ) );
     DataFlexSizer->Add( m_CoCopyButton, 0, wxTOP|wxRIGHT|wxLEFT, 5 );
 
     wxStaticText * CoStaticText;
-    CoStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Composer")), wxDefaultPosition, wxDefaultSize, 0 );
+    CoStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Composer")), wxDefaultPosition, wxDefaultSize, 0 );
     CoStaticText->Wrap( -1 );
     DataFlexSizer->Add( CoStaticText, 0, wxALIGN_RIGHT|wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
-    m_CompComboBox = new wxComboBox( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
+    m_CompComboBox = new wxComboBox( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
     DataFlexSizer->Add( m_CompComboBox, 1, wxEXPAND|wxTOP|wxRIGHT, 5 );
 
     // Comment
-    m_CommentCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_CommentCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_CommentCopyButton->SetToolTip( _( "Copy the comment to all the tracks you are editing" ) );
     DataFlexSizer->Add( m_CommentCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
     wxStaticText * CommentStaticText;
-    CommentStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Comment")), wxDefaultPosition, wxDefaultSize, 0 );
+    CommentStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Comment")), wxDefaultPosition, wxDefaultSize, 0 );
     CommentStaticText->Wrap( -1 );
     DataFlexSizer->Add( CommentStaticText, 0, wxALIGN_RIGHT|wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
-    m_CommentText = new wxTextCtrl( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxSize( -1, 54 ), wxTE_MULTILINE );
+    m_CommentText = new wxTextCtrl( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxSize( -1, 54 ), wxTE_MULTILINE );
     DataFlexSizer->Add( m_CommentText, 1, wxEXPAND|wxTOP|wxRIGHT, 5 );
 
     // Number
-    m_NuCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_NuCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_NuCopyButton->SetToolTip( _( "Copy the number to all the tracks you are editing" ) );
     DataFlexSizer->Add( m_NuCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
-    NuStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Number")), wxDefaultPosition, wxDefaultSize, 0 );
+    NuStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Number")), wxDefaultPosition, wxDefaultSize, 0 );
     NuStaticText->Wrap( -1 );
     DataFlexSizer->Add( NuStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT|wxTOP|wxRIGHT, 5 );
 
     wxBoxSizer* DiskSizer;
     DiskSizer = new wxBoxSizer( wxHORIZONTAL );
 
-    m_NumberTextCtrl = new wxTextCtrl( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
+    m_NumberTextCtrl = new wxTextCtrl( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
     DiskSizer->Add( m_NumberTextCtrl, 0, wxALIGN_CENTER_VERTICAL|wxTOP, 5 );
 
-    m_NuOrderButton= new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_numerate ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_NuOrderButton= new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_numerate ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_NuOrderButton->SetToolTip( _( "Enumerate the tracks in the order they were added for editing" ) );
     DiskSizer->Add( m_NuOrderButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT, 5 );
 
     DiskSizer->Add( 0, 0, 1, wxEXPAND, 5 );
 
     // Disk
-    m_DiCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_DiCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_DiCopyButton->SetToolTip( _( "Copy the disk to all the tracks you are editing" ) );
     DiskSizer->Add( m_DiCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
     wxStaticText * DiStaticText;
-    DiStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Disk")), wxDefaultPosition, wxDefaultSize, 0 );
+    DiStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Disk")), wxDefaultPosition, wxDefaultSize, 0 );
     DiStaticText->Wrap( -1 );
     DiskSizer->Add( DiStaticText, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT, 5 );
 
-    m_DiskTextCtrl = new wxTextCtrl( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
+    m_DiskTextCtrl = new wxTextCtrl( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
     DiskSizer->Add( m_DiskTextCtrl, 0, wxTOP|wxRIGHT, 5 );
 
     DataFlexSizer->Add( DiskSizer, 1, wxEXPAND, 5 );
 
     // Genre
-    m_GeCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_GeCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_GeCopyButton->SetToolTip( _( "Copy the genre name to all songs you are editing" ) );
     DataFlexSizer->Add( m_GeCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
-    GeStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Genre")), wxDefaultPosition, wxDefaultSize, 0 );
+    GeStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Genre")), wxDefaultPosition, wxDefaultSize, 0 );
     GeStaticText->Wrap( -1 );
     DataFlexSizer->Add( GeStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT|wxTOP|wxRIGHT, 5 );
 
-    m_GenreComboBox = new wxComboBox( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
+    m_GenreComboBox = new wxComboBox( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, DummyArray, wxCB_DROPDOWN );
     DataFlexSizer->Add( m_GenreComboBox, 1, wxEXPAND|wxTOP|wxRIGHT|wxALIGN_CENTER_VERTICAL, 5 );
 
     // Year
-    m_YeCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_YeCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_YeCopyButton->SetToolTip( _( "Copy the year to all songs you are editing" ) );
     DataFlexSizer->Add( m_YeCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
-    YeStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Year")), wxDefaultPosition, wxDefaultSize, 0 );
+    YeStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Year")), wxDefaultPosition, wxDefaultSize, 0 );
     YeStaticText->Wrap( -1 );
     DataFlexSizer->Add( YeStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT|wxTOP|wxRIGHT, 5 );
 
-    m_YearTextCtrl = new wxTextCtrl( DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
+    m_YearTextCtrl = new wxTextCtrl( m_DetailPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
 
     // Rating
     wxBoxSizer * RatingSizer;
@@ -286,15 +284,15 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     RatingSizer->Add( m_YearTextCtrl, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxBOTTOM|wxRIGHT, 5 );
     RatingSizer->Add( 0, 0, 1, wxEXPAND, 5 );
 
-    m_RaCopyButton = new wxBitmapButton( DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_RaCopyButton = new wxBitmapButton( m_DetailPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     RatingSizer->Add( m_RaCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxALL, 5 );
 
     wxStaticText * RaStaticText;
-    RaStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s:", _("Rating")), wxDefaultPosition, wxDefaultSize, 0 );
+    RaStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s:", _("Rating")), wxDefaultPosition, wxDefaultSize, 0 );
     RaStaticText->Wrap( -1 );
     RatingSizer->Add( RaStaticText, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxBOTTOM|wxRIGHT, 5 );
 
-    m_Rating = new guRating( DetailPanel, GURATING_STYLE_BIG );
+    m_Rating = new guRating( m_DetailPanel, GURATING_STYLE_BIG );
     RatingSizer->Add( m_Rating, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxBOTTOM|wxRIGHT, 5 );
 
     DataFlexSizer->Add( RatingSizer, 1, wxEXPAND, 5 );
@@ -305,32 +303,32 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     wxBoxSizer* MoreDetailsSizer;
     MoreDetailsSizer = new wxBoxSizer( wxHORIZONTAL );
 
-    m_DetailLeftInfoStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s\t:\n%s\t:", _("File Type"), _("Length")), wxDefaultPosition, wxDefaultSize, 0 );
+    m_DetailLeftInfoStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s\t:\n%s\t:", _("File Type"), _("Length")), wxDefaultPosition, wxDefaultSize, 0 );
     m_DetailLeftInfoStaticText->Wrap( -1 );
     MoreDetailsSizer->Add( m_DetailLeftInfoStaticText, 1, wxALL|wxEXPAND, 5 );
 
-    m_DetailRightInfoStaticText = new wxStaticText( DetailPanel, wxID_ANY, wxString::Format("%s\t:\n%s\t:", _("Bit Rate"), _("File Size")), wxDefaultPosition, wxDefaultSize, 0 );
+    m_DetailRightInfoStaticText = new wxStaticText( m_DetailPanel, wxID_ANY, wxString::Format("%s\t:\n%s\t:", _("Bit Rate"), _("File Size")), wxDefaultPosition, wxDefaultSize, 0 );
     m_DetailRightInfoStaticText->Wrap( -1 );
     MoreDetailsSizer->Add( m_DetailRightInfoStaticText, 1, wxALL|wxEXPAND, 5 );
 
     MainDetailSizer->Add( MoreDetailsSizer, 1, wxEXPAND, 5 );
 
-    DetailPanel->SetSizer( MainDetailSizer );
-    DetailPanel->Layout();
-    DataFlexSizer->Fit( DetailPanel );
-    m_MainNotebook->AddPage( DetailPanel, _( "Details" ), true );
+    m_DetailPanel->SetSizer( MainDetailSizer );
+    m_DetailPanel->Layout();
+    DataFlexSizer->Fit( m_DetailPanel );
+    m_MainNotebook->AddPage( m_DetailPanel, _( "Details" ), true );
 
     //
     // Pictures
     //
-    PicturePanel = new wxPanel( m_MainNotebook, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxTAB_TRAVERSAL );
+    m_PicturePanel = new wxPanel( m_MainNotebook, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxTAB_TRAVERSAL );
     wxBoxSizer* PictureSizer;
     PictureSizer = new wxBoxSizer( wxVERTICAL );
 
     wxStaticBoxSizer* PictureBitmapSizer;
-    PictureBitmapSizer = new wxStaticBoxSizer( new wxStaticBox( PicturePanel, wxID_ANY, wxEmptyString ), wxVERTICAL );
+    PictureBitmapSizer = new wxStaticBoxSizer( new wxStaticBox( m_PicturePanel, wxID_ANY, wxEmptyString ), wxVERTICAL );
 
-    m_PictureBitmap = new wxStaticBitmap( PicturePanel, wxID_ANY, wxNullBitmap, wxDefaultPosition, wxSize( 250,250 ), wxSUNKEN_BORDER );
+    m_PictureBitmap = new wxStaticBitmap( m_PicturePanel, wxID_ANY, wxNullBitmap, wxDefaultPosition, wxSize( 250,250 ), wxSUNKEN_BORDER );
     PictureBitmapSizer->Add( m_PictureBitmap, 0, wxALL|wxALIGN_CENTER_VERTICAL|wxALIGN_CENTER_HORIZONTAL, 5 );
 
     PictureSizer->Add( PictureBitmapSizer, 0, wxALIGN_CENTER_HORIZONTAL, 5 );
@@ -338,19 +336,19 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     wxBoxSizer* PictureButtonSizer;
     PictureButtonSizer = new wxBoxSizer( wxHORIZONTAL );
 
-    m_AddPicButton = new wxBitmapButton( PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_add ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_AddPicButton = new wxBitmapButton( m_PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_add ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_AddPicButton->SetToolTip( _( "Add a picture from file to the current track" ) );
     PictureButtonSizer->Add( m_AddPicButton, 0, wxALL|wxALIGN_CENTER_VERTICAL, 5 );
 
-    m_DelPicButton = new wxBitmapButton( PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_del ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_DelPicButton = new wxBitmapButton( m_PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_del ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_DelPicButton->SetToolTip( _( "Delete the picture from the current track" ) );
     PictureButtonSizer->Add( m_DelPicButton, 0, wxALL|wxALIGN_CENTER_VERTICAL, 5 );
 
-    m_SavePicButton = new wxBitmapButton( PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_doc_save ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_SavePicButton = new wxBitmapButton( m_PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_doc_save ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_SavePicButton->SetToolTip( _( "Save the current picture to file" ) );
     PictureButtonSizer->Add( m_SavePicButton, 0, wxALL|wxALIGN_CENTER_VERTICAL, 5 );
 
-    m_SearchPicButton = new wxBitmapButton( PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_search ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_SearchPicButton = new wxBitmapButton( m_PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_search ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_SearchPicButton->SetToolTip( _( "Search the album cover" ) );
     PictureButtonSizer->Add( m_SearchPicButton, 0, wxALL|wxALIGN_CENTER_VERTICAL, 5 );
 
@@ -359,59 +357,59 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
 
     PictureButtonSizer->Add( 10, 0, 0, wxEXPAND, 5 );
 
-    m_CopyPicButton = new wxBitmapButton( PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_CopyPicButton = new wxBitmapButton( m_PicturePanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_CopyPicButton->SetToolTip( _( "Copy the current picture to all the tracks you are editing" ) );
     PictureButtonSizer->Add( m_CopyPicButton, 0, wxALL, 5 );
 
     PictureSizer->Add( PictureButtonSizer, 0, wxALIGN_CENTER_HORIZONTAL, 5 );
 
-    PicturePanel->SetSizer( PictureSizer );
-    PicturePanel->Layout();
-    PictureSizer->Fit( PicturePanel );
-    m_MainNotebook->AddPage( PicturePanel, _( "Pictures" ), false );
+    m_PicturePanel->SetSizer( PictureSizer );
+    m_PicturePanel->Layout();
+    PictureSizer->Fit( m_PicturePanel );
+    m_MainNotebook->AddPage( m_PicturePanel, _( "Pictures" ), false );
 
     //
     // Lyrics
     //
-    wxPanel * LyricsPanel = new wxPanel( m_MainNotebook, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxTAB_TRAVERSAL );
+    m_LyricsPanel = new wxPanel( m_MainNotebook, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxTAB_TRAVERSAL );
     wxBoxSizer * LyricsSizer;
     LyricsSizer = new wxBoxSizer( wxVERTICAL );
 
     wxBoxSizer* LyricsTopSizer;
     LyricsTopSizer = new wxBoxSizer( wxHORIZONTAL );
 
-    wxStaticText * ArtistStaticText = new wxStaticText( LyricsPanel, wxID_ANY, wxString::Format("%s:", _("Artist")), wxDefaultPosition, wxDefaultSize, 0 );
+    wxStaticText * ArtistStaticText = new wxStaticText( m_LyricsPanel, wxID_ANY, wxString::Format("%s:", _("Artist")), wxDefaultPosition, wxDefaultSize, 0 );
     ArtistStaticText->Wrap( -1 );
     LyricsTopSizer->Add( ArtistStaticText, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxLEFT, 5 );
 
-    m_LyricArtistTextCtrl = new wxTextCtrl( LyricsPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
+    m_LyricArtistTextCtrl = new wxTextCtrl( m_LyricsPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
     LyricsTopSizer->Add( m_LyricArtistTextCtrl, 1, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT, 5 );
 
-    wxStaticText * TrackStaticText = new wxStaticText( LyricsPanel, wxID_ANY, wxString::Format("%s:", _("Track")), wxDefaultPosition, wxDefaultSize, 0 );
+    wxStaticText * TrackStaticText = new wxStaticText( m_LyricsPanel, wxID_ANY, wxString::Format("%s:", _("Track")), wxDefaultPosition, wxDefaultSize, 0 );
     TrackStaticText->Wrap( -1 );
     LyricsTopSizer->Add( TrackStaticText, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxLEFT, 5 );
 
-    m_LyricTrackTextCtrl = new wxTextCtrl( LyricsPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
+    m_LyricTrackTextCtrl = new wxTextCtrl( m_LyricsPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
     LyricsTopSizer->Add( m_LyricTrackTextCtrl, 1, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT, 5 );
 
-    m_LyricReloadButton = new wxBitmapButton( LyricsPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_search_again ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_LyricReloadButton = new wxBitmapButton( m_LyricsPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_search_again ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_LyricReloadButton->SetToolTip( _( "Search for lyrics" ) );
     LyricsTopSizer->Add( m_LyricReloadButton, 0, wxALIGN_CENTER_VERTICAL|wxTOP|wxRIGHT|wxLEFT, 5 );
 
     LyricsSizer->Add( LyricsTopSizer, 0, wxEXPAND, 5 );
 
-    m_LyricsTextCtrl = new wxTextCtrl( LyricsPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_CENTRE|wxTE_DONTWRAP|wxTE_MULTILINE );
+    m_LyricsTextCtrl = new wxTextCtrl( m_LyricsPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_CENTRE|wxTE_DONTWRAP|wxTE_MULTILINE );
     LyricsSizer->Add( m_LyricsTextCtrl, 1, wxALL|wxEXPAND, 5 );
 
-    LyricsPanel->SetSizer( LyricsSizer );
-    LyricsPanel->Layout();
-    LyricsSizer->Fit( LyricsPanel );
-    m_MainNotebook->AddPage( LyricsPanel, _( "Lyrics" ), false );
+    m_LyricsPanel->SetSizer( LyricsSizer );
+    m_LyricsPanel->Layout();
+    LyricsSizer->Fit( m_LyricsPanel );
+    m_MainNotebook->AddPage( m_LyricsPanel, _( "Lyrics" ), false );
 
     //
     // MusicBrainz
     //
-    MBrainzPanel = new wxPanel( m_MainNotebook, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxTAB_TRAVERSAL );
+    m_MBrainzPanel = new wxPanel( m_MainNotebook, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxTAB_TRAVERSAL );
     wxBoxSizer* MBrainzSizer;
     MBrainzSizer = new wxBoxSizer( wxVERTICAL );
 
@@ -420,23 +418,23 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
 
     wxStaticText * MBQueryArtistStaticText;
     wxStaticText * MBQueryTitleStaticText;
-    MBQueryArtistStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Artist")), wxDefaultPosition, wxDefaultSize, 0 );
+    MBQueryArtistStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Artist")), wxDefaultPosition, wxDefaultSize, 0 );
     MBQueryArtistStaticText->Wrap( -1 );
     MBQuerySizer->Add( MBQueryArtistStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALL, 5 );
 
-    m_MBQueryArtistTextCtrl = new wxTextCtrl( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
+    m_MBQueryArtistTextCtrl = new wxTextCtrl( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
     m_MBQueryArtistTextCtrl->SetToolTip( _( "Type the artist name to search in musicbrainz" ) );
     MBQuerySizer->Add( m_MBQueryArtistTextCtrl, 1, wxALIGN_CENTER_VERTICAL|wxTOP|wxBOTTOM|wxRIGHT, 5 );
 
-    MBQueryTitleStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Title")), wxDefaultPosition, wxDefaultSize, 0 );
+    MBQueryTitleStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Title")), wxDefaultPosition, wxDefaultSize, 0 );
     MBQueryTitleStaticText->Wrap( -1 );
     MBQuerySizer->Add( MBQueryTitleStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALL, 5 );
 
-    m_MBQueryTitleTextCtrl = new wxTextCtrl( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
+    m_MBQueryTitleTextCtrl = new wxTextCtrl( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, 0 );
     m_MBQueryTitleTextCtrl->SetToolTip( _( "Type the album name to search in musicbrainz" ) );
     MBQuerySizer->Add( m_MBQueryTitleTextCtrl, 1, wxALIGN_CENTER_VERTICAL|wxTOP|wxBOTTOM|wxRIGHT, 5 );
 
-    m_MBQueryClearButton = new wxBitmapButton( MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_clear ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_MBQueryClearButton = new wxBitmapButton( m_MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_clear ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_MBQueryClearButton->SetToolTip( _( "Clear the search fields so it search using the music fingerprint" ) );
     MBQuerySizer->Add( m_MBQueryClearButton, 0, wxTOP|wxBOTTOM|wxRIGHT, 5 );
 
@@ -446,24 +444,24 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     wxBoxSizer* MBrainzTopSizer;
     MBrainzTopSizer = new wxBoxSizer( wxHORIZONTAL );
 
-    MBAlbumStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Album")), wxDefaultPosition, wxDefaultSize, 0 );
+    MBAlbumStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Album")), wxDefaultPosition, wxDefaultSize, 0 );
     MBAlbumStaticText->Wrap( -1 );
     MBrainzTopSizer->Add( MBAlbumStaticText, 0, wxALIGN_CENTER_VERTICAL|wxALL, 5 );
 
     wxArrayString m_MBAlbumChoiceChoices;
-    m_MBAlbumChoice = new wxChoice( MBrainzPanel, wxID_ANY, wxDefaultPosition, wxDefaultSize, m_MBAlbumChoiceChoices, 0 );
+    m_MBAlbumChoice = new wxChoice( m_MBrainzPanel, wxID_ANY, wxDefaultPosition, wxDefaultSize, m_MBAlbumChoiceChoices, 0 );
     m_MBAlbumChoice->SetToolTip( _( "Select the album found in musicbrainz" ) );
     m_MBAlbumChoice->SetSelection( 0 );
 
     MBrainzTopSizer->Add( m_MBAlbumChoice, 1, wxALIGN_CENTER_VERTICAL|wxTOP|wxBOTTOM|wxRIGHT, 5 );
 
-    m_MBAddButton = new wxBitmapButton( MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_search ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_MBAddButton = new wxBitmapButton( m_MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_search ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_MBAddButton->SetToolTip( _( "Search albums in musicbrainz" ) );
     //m_MBAddButton->Enable( false );
 
     MBrainzTopSizer->Add( m_MBAddButton, 0, wxTOP|wxBOTTOM|wxRIGHT|wxALIGN_CENTER_VERTICAL, 5 );
 
-    m_MBCopyButton = new wxBitmapButton( MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_MBCopyButton = new wxBitmapButton( m_MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_MBCopyButton->SetToolTip( _( "Copy the content of the album to the edited tracks" ) );
     m_MBCopyButton->Enable( false );
 
@@ -473,7 +471,7 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
 
     // Details
     wxStaticBoxSizer* MBDetailSizer;
-    MBDetailSizer = new wxStaticBoxSizer( new wxStaticBox( MBrainzPanel, wxID_ANY, wxString::Format(" %s ", _("Details")) ), wxVERTICAL );
+    MBDetailSizer = new wxStaticBoxSizer( new wxStaticBox( m_MBrainzPanel, wxID_ANY, wxString::Format(" %s ", _("Details")) ), wxVERTICAL );
 
     wxFlexGridSizer* MBDetailFlexGridSizer;
     MBDetailFlexGridSizer = new wxFlexGridSizer( 3, 0, 0 );
@@ -482,106 +480,106 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     MBDetailFlexGridSizer->SetNonFlexibleGrowMode( wxFLEX_GROWMODE_SPECIFIED );
 
     // Title
-    m_MBTitleStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Title")), wxDefaultPosition, wxDefaultSize, 0 );
+    m_MBTitleStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Title")), wxDefaultPosition, wxDefaultSize, 0 );
     m_MBTitleStaticText->Wrap( -1 );
     MBDetailFlexGridSizer->Add( m_MBTitleStaticText, 0, wxBOTTOM|wxRIGHT|wxLEFT|wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT, 5 );
 
-    m_MBTitleTextCtrl = new wxTextCtrl( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
+    m_MBTitleTextCtrl = new wxTextCtrl( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
     MBDetailFlexGridSizer->Add( m_MBTitleTextCtrl, 1, wxEXPAND|wxBOTTOM|wxRIGHT, 5 );
 
-    m_MBTiCopyButton = new wxBitmapButton( MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_MBTiCopyButton = new wxBitmapButton( m_MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_MBTiCopyButton->SetToolTip( _( "Copy the song names to the edited tracks" ) );
     MBDetailFlexGridSizer->Add( m_MBTiCopyButton, 0, wxBOTTOM|wxRIGHT|wxALIGN_CENTER_VERTICAL, 5 );
 
     // Artist
-    m_MBArtistStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Artist")), wxDefaultPosition, wxDefaultSize, 0 );
+    m_MBArtistStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Artist")), wxDefaultPosition, wxDefaultSize, 0 );
     m_MBArtistStaticText->Wrap( -1 );
     MBDetailFlexGridSizer->Add( m_MBArtistStaticText, 0, wxALIGN_RIGHT|wxALIGN_CENTER_VERTICAL|wxBOTTOM|wxRIGHT|wxLEFT, 5 );
 
-    m_MBArtistTextCtrl = new wxTextCtrl( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
+    m_MBArtistTextCtrl = new wxTextCtrl( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
     MBDetailFlexGridSizer->Add( m_MBArtistTextCtrl, 1, wxEXPAND|wxBOTTOM|wxRIGHT, 5 );
 
-    m_MBArCopyButton = new wxBitmapButton( MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_MBArCopyButton = new wxBitmapButton( m_MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_MBArCopyButton->SetToolTip( _( "Copy the artist to the edited tracks" ) );
     MBDetailFlexGridSizer->Add( m_MBArCopyButton, 0, wxBOTTOM|wxRIGHT|wxALIGN_CENTER_VERTICAL, 5 );
 
     // Album Artist
-    m_MBAlbumArtistStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("A. Artist")), wxDefaultPosition, wxDefaultSize, 0 );
+    m_MBAlbumArtistStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("A. Artist")), wxDefaultPosition, wxDefaultSize, 0 );
     m_MBAlbumArtistStaticText->Wrap( -1 );
     MBDetailFlexGridSizer->Add( m_MBAlbumArtistStaticText, 0, wxALIGN_RIGHT|wxALIGN_CENTER_VERTICAL|wxBOTTOM|wxRIGHT|wxLEFT, 5 );
 
-    m_MBAlbumArtistTextCtrl = new wxTextCtrl( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
+    m_MBAlbumArtistTextCtrl = new wxTextCtrl( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
     MBDetailFlexGridSizer->Add( m_MBAlbumArtistTextCtrl, 1, wxEXPAND|wxBOTTOM|wxRIGHT, 5 );
 
-    m_MBAlArCopyButton = new wxBitmapButton( MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_MBAlArCopyButton = new wxBitmapButton( m_MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_MBAlArCopyButton->SetToolTip( _( "Copy the album artist to the edited tracks" ) );
     MBDetailFlexGridSizer->Add( m_MBAlArCopyButton, 0, wxBOTTOM|wxRIGHT|wxALIGN_CENTER_VERTICAL, 5 );
 
     // Album
-    m_MBAlbumStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Album")), wxDefaultPosition, wxDefaultSize, 0 );
+    m_MBAlbumStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Album")), wxDefaultPosition, wxDefaultSize, 0 );
     m_MBAlbumStaticText->Wrap( -1 );
     MBDetailFlexGridSizer->Add( m_MBAlbumStaticText, 0, wxBOTTOM|wxRIGHT|wxLEFT|wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT, 5 );
 
-    m_MBAlbumTextCtrl = new wxTextCtrl( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
+    m_MBAlbumTextCtrl = new wxTextCtrl( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
     MBDetailFlexGridSizer->Add( m_MBAlbumTextCtrl, 1, wxEXPAND|wxBOTTOM|wxRIGHT, 5 );
 
-    m_MBAlCopyButton = new wxBitmapButton( MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_MBAlCopyButton = new wxBitmapButton( m_MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_MBAlCopyButton->SetToolTip( _( "Copy the album to the edited tracks" ) );
     MBDetailFlexGridSizer->Add( m_MBAlCopyButton, 0, wxBOTTOM|wxRIGHT|wxALIGN_CENTER_VERTICAL, 5 );
 
     // Year
-    m_MBYearStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Year")), wxDefaultPosition, wxDefaultSize, 0 );
+    m_MBYearStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Year")), wxDefaultPosition, wxDefaultSize, 0 );
     m_MBYearStaticText->Wrap( -1 );
     MBDetailFlexGridSizer->Add( m_MBYearStaticText, 0, wxBOTTOM|wxRIGHT|wxLEFT|wxALIGN_CENTER_VERTICAL|wxALIGN_RIGHT, 5 );
 
-    m_MBYearTextCtrl = new wxTextCtrl( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
+    m_MBYearTextCtrl = new wxTextCtrl( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
     MBDetailFlexGridSizer->Add( m_MBYearTextCtrl, 0, wxEXPAND|wxBOTTOM|wxRIGHT, 5 );
 
-    m_MBDaCopyButton = new wxBitmapButton( MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_MBDaCopyButton = new wxBitmapButton( m_MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_MBDaCopyButton->SetToolTip( _( "Copy the date to the edited tracks" ) );
     MBDetailFlexGridSizer->Add( m_MBDaCopyButton, 0, wxBOTTOM|wxRIGHT|wxALIGN_CENTER_VERTICAL, 5 );
 
     // Size and number
-    m_MBLengthStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, _("Length:"), wxDefaultPosition, wxDefaultSize, 0 );
+    m_MBLengthStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, _("Length:"), wxDefaultPosition, wxDefaultSize, 0 );
     m_MBLengthStaticText->Wrap( -1 );
     MBDetailFlexGridSizer->Add( m_MBLengthStaticText, 0, wxBOTTOM|wxRIGHT|wxLEFT|wxALIGN_CENTER_VERTICAL, 5 );
 
     wxBoxSizer* MBNumberSizer;
     MBNumberSizer = new wxBoxSizer( wxHORIZONTAL );
 
-    m_MBLengthTextCtrl = new wxTextCtrl( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
+    m_MBLengthTextCtrl = new wxTextCtrl( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
     MBNumberSizer->Add( m_MBLengthTextCtrl, 0, wxALIGN_CENTER_VERTICAL|wxBOTTOM|wxRIGHT, 5 );
 
     MBNumberSizer->Add( 0, 0, 1, wxEXPAND, 5 );
 
-    m_MBNumberStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Number")), wxDefaultPosition, wxDefaultSize, 0 );
+    m_MBNumberStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxString::Format("%s:", _("Number")), wxDefaultPosition, wxDefaultSize, 0 );
     m_MBNumberStaticText->Wrap( -1 );
     MBNumberSizer->Add( m_MBNumberStaticText, 0, wxALIGN_CENTER_VERTICAL|wxBOTTOM|wxRIGHT|wxLEFT, 5 );
 
-    m_MBNumberTextCtrl = new wxTextCtrl( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
+    m_MBNumberTextCtrl = new wxTextCtrl( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxTE_READONLY );
     MBNumberSizer->Add( m_MBNumberTextCtrl, 0, wxBOTTOM|wxRIGHT|wxALIGN_CENTER_VERTICAL, 5 );
 
     MBDetailFlexGridSizer->Add( MBNumberSizer, 1, wxEXPAND, 5 );
 
-    m_MBNuCopyButton = new wxBitmapButton( MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
+    m_MBNuCopyButton = new wxBitmapButton( m_MBrainzPanel, wxID_ANY, guImage( guIMAGE_INDEX_tiny_edit_copy ), wxDefaultPosition, wxDefaultSize, wxBU_AUTODRAW );
     m_MBNuCopyButton->SetToolTip( _( "Copy the number to the edited tracks" ) );
     MBDetailFlexGridSizer->Add( m_MBNuCopyButton, 0, wxALIGN_CENTER_VERTICAL|wxBOTTOM|wxRIGHT, 5 );
 
     MBDetailSizer->Add( MBDetailFlexGridSizer, 0, wxEXPAND, 5 );
 
-    MBrainzStaticLine = new wxStaticLine( MBrainzPanel, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxLI_HORIZONTAL );
+    MBrainzStaticLine = new wxStaticLine( m_MBrainzPanel, wxID_ANY, wxDefaultPosition, wxDefaultSize, wxLI_HORIZONTAL );
     MBDetailSizer->Add( MBrainzStaticLine, 0, wxEXPAND | wxALL, 5 );
 
-    m_MBInfoStaticText = new wxStaticText( MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxST_NO_AUTORESIZE );
+    m_MBInfoStaticText = new wxStaticText( m_MBrainzPanel, wxID_ANY, wxEmptyString, wxDefaultPosition, wxDefaultSize, wxST_NO_AUTORESIZE );
     m_MBInfoStaticText->Wrap( 398 );
     MBDetailSizer->Add( m_MBInfoStaticText, 1, wxALL|wxEXPAND, 5 );
 
     MBrainzSizer->Add( MBDetailSizer, 1, wxEXPAND|wxBOTTOM|wxRIGHT|wxLEFT, 5 );
 
-    MBrainzPanel->SetSizer( MBrainzSizer );
-    MBrainzPanel->Layout();
-    MBrainzSizer->Fit( MBrainzPanel );
-    m_MainNotebook->AddPage( MBrainzPanel, wxT( "MusicBrainz" ), false );
+    m_MBrainzPanel->SetSizer( MBrainzSizer );
+    m_MBrainzPanel->Layout();
+    MBrainzSizer->Fit( m_MBrainzPanel );
+    m_MainNotebook->AddPage( m_MBrainzPanel, wxT( "MusicBrainz" ), false );
 
     DetailSizer->Add( m_MainNotebook, 1, wxEXPAND | wxALL, 5 );
     // End MusicBrainz
@@ -636,9 +634,10 @@ guTrackEditor::guTrackEditor( wxWindow * parent, guDbLibrary * db, guTrackArray 
     {
         ItemsText.Add( ( * m_Items )[ index ].m_FileName );
         // Fill the initial Images of the files
-        m_Images->Add( guTagGetPicture( ( * m_Items )[ index ].m_FileName ) );
-        m_Lyrics->Add( guTagGetLyrics( ( * m_Items )[ index ].m_FileName ) );
-        m_ChangedFlags->Add( guTRACK_CHANGED_DATA_NONE );
+        m_Images->Add(guTagGetPicture((*m_Items)[index].m_FileName));
+        m_Lyrics->Add(guTagGetLyrics((*m_Items)[index].m_FileName));
+        m_ChangedFlags->Add(guTRACK_CHANGED_DATA_NONE);
+        m_CanEdit.Add(guTagGetCanEdit((*m_Items)[index].m_FileName) ? 1 : 0);
     }
     m_SongListBox->InsertItems( ItemsText, 0 );
     m_SongListBox->SetFocus();
@@ -853,15 +852,19 @@ void guTrackEditor::OnMoveUpBtnClick( wxCommandEvent &event )
     wxImage * MovedImage = ( * m_Images )[ m_CurItem ];
     wxString MovedLyric = ( * m_Lyrics )[ m_CurItem ];
     int MovedFlag = ( * m_ChangedFlags )[ m_CurItem ];
+    short CanEdit = (m_CanEdit)[m_CurItem];
+
     m_Images->RemoveAt( m_CurItem );
     m_Lyrics->RemoveAt( m_CurItem );
     m_ChangedFlags->RemoveAt( m_CurItem );
+    m_CanEdit.RemoveAt(m_CurItem);
     m_SongListBox->SetString( m_CurItem, m_SongListBox->GetString( m_CurItem - 1 ) );
     m_CurItem--;
     m_Items->Insert( MovedTrack, m_CurItem );
     m_Images->Insert( MovedImage, m_CurItem );
     m_Lyrics->Insert( MovedLyric, m_CurItem );
     m_ChangedFlags->Insert( MovedFlag, m_CurItem );
+    m_CanEdit.Insert(CanEdit, m_CurItem);
     m_SongListBox->SetString( m_CurItem, FileName );
 
     m_SongListBox->SetSelection( m_CurItem );
@@ -878,15 +881,19 @@ void guTrackEditor::OnMoveDownBtnClick( wxCommandEvent &event )
     wxImage * MovedImage = ( * m_Images )[ m_CurItem ];
     wxString MovedLyric = ( * m_Lyrics )[ m_CurItem ];
     int MovedFlag = ( * m_ChangedFlags )[ m_CurItem ];
+    short CanEdit = (m_CanEdit)[m_CurItem];
+
     m_Images->RemoveAt( m_CurItem );
     m_Lyrics->RemoveAt( m_CurItem );
     m_ChangedFlags->RemoveAt( m_CurItem );
+    m_CanEdit.RemoveAt( m_CurItem );
     m_SongListBox->SetString( m_CurItem, m_SongListBox->GetString( m_CurItem + 1 ) );
     m_CurItem++;
     m_Items->Insert( MovedTrack, m_CurItem );
     m_Images->Insert( MovedImage, m_CurItem );
     m_Lyrics->Insert( MovedLyric, m_CurItem );
     m_ChangedFlags->Insert( MovedFlag, m_CurItem );
+    m_CanEdit.Insert(CanEdit, m_CurItem);
     m_SongListBox->SetString( m_CurItem, FileName );
 
     m_SongListBox->SetSelection( m_CurItem );
@@ -946,6 +953,8 @@ void guTrackEditor::ReadItemData()
         m_LyricArtistTextCtrl->SetValue( Track->m_ArtistName );
         m_LyricTrackTextCtrl->SetValue( Track->m_SongName );
         m_LyricsTextCtrl->SetValue( ( * m_Lyrics )[ m_CurItem ] );
+
+        EnableEdit((m_CanEdit)[m_CurItem] > 0);
     }
     else
     {
@@ -987,6 +996,16 @@ void guTrackEditor::ReadItemData()
     UpdateMBrainzTrackInfo();
 }
 
+
+// -------------------------------------------------------------------------------- //
+void guTrackEditor::EnableEdit(const bool is_enabled) const
+{
+    m_DetailPanel->Enable(is_enabled);
+    m_LyricsPanel->Enable(is_enabled);
+    m_PicturePanel->Enable(is_enabled);
+    m_MBrainzPanel->Enable(is_enabled);
+}
+
 // -------------------------------------------------------------------------------- //
 void guTrackEditor::SetTagField( wxString &field, const wxString &newval, int &changedflags, const int flagval )
 {
@@ -1010,7 +1029,7 @@ void guTrackEditor::SetTagField( int &field, const int newval, int &changedflags
 }
 
 // -------------------------------------------------------------------------------- //
-void guTrackEditor::WriteItemData( void )
+void guTrackEditor::WriteItemData()
 {
     //guLogMessage( wxT( "WriteItemData: %i" ), m_CurItem );
     if( m_CurItem >= 0 )